Relocation of API methods in separated modules is completed. All methods are functional in its primary data target, some secondary data gathering is yet to be implemented. Redesign of the public API interface is completed.
